feat: Add support for perf test the new tracker importer
This adds support for testing the new tracker importer.
This commit is configured to perf test events on a loop of create with 5 events.
It is trivial to modify to test only Tei or Enrollment inserts or any combination.

This test needs a DHIS2 instance running against an empty DHIS2 database with standard user/pass admin:district

The Main.class first tries to insert metadata.
Then it generates n (5) number of tei,enrollments random id pairs.
This map of ids is the input for the Tei, Enrollment and Event generators.